# Cache struct lookup tables